+++ CHANGELOG.md
@@ -88,3 +88,6 @@ Ref: https://keepachangelog.com/en/1.0.0/
     * `GET /v0/org/:org_id/roles/:role_id/keys`
 * (btc) Implemented API methods: 
     * `POST /v0/org/:org_id/btc/taproot/sign/:pubkey` 
+
+### Improvements
+* (client) Refactor logging for requests with body